kern: don't allocate stack for FPU on armv6m
This takes 72 bytes off the stack usage for every task, since the kernel
sets the initial task SP based on the assumption that the processor will
consume an exception frame from it. We were overestimating the size of
the exception frame by treating it as invariant across v6m and v7/8m.